+++ decidim-proposals/db/migrate/20190416112905_add_collaborative_draft_origin_to_proposal.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class AddCollaborativeDraftOriginToProposal < ActiveRecord::Migration[5.2]
+  def change
+    add_column :decidim_proposals_proposals, :collaborative_draft_origin, :boolean, default: false
+  end
+end