+++ decidim-proposals/spec/services/decidim/proposals/proposal_serializer_spec.rb
@@ -4,7 +4,7 @@ require "spec_helper"
 
 module Decidim
   module Proposals
-    describe ProposalSerializer do
+    describe ProposalSerializer, processing_uploads_for: Decidim::AttachmentUploader do
       subject do
         described_class.new(proposal)
       end
@@ -18,10 +18,24 @@ module Decidim
       let!(:meetings_component) { create(:component, manifest_name: "meetings", participatory_space: participatory_process) }
       let(:meetings) { create_list(:meeting, 2, component: meetings_component) }
 
-      let!(:proposals_component) { create(:component, manifest_name: "proposals", participatory_space: participatory_process) }
+      let!(:proposals_component) { create(:proposal_component, :with_attachments_allowed, manifest_name: "proposals", participatory_space: participatory_process) }
       let(:other_proposals) { create_list(:proposal, 2, component: proposals_component) }
 
+      let!(:emendation) { create(:proposal, component: component) }
+      let!(:amendment) { create(:amendment, amendable: proposal, emendation: emendation) }
+
+      let!(:attachment) { Decidim::Attachment.new(attachment_params) }
+      let(:attachment_params) do
+        {
+          title: "My attachment",
+          file: Decidim::Dev.test_file("city.jpeg", "image/jpeg"),
+          content_type: "image/jpeg",
+          attached_to: proposal
+        }
+      end
+
       before do
+        attachment.save!
         proposal.update!(category: category)
         proposal.update!(scope: scope)
         proposal.link_resources(meetings, "proposals_from_meeting")
@@ -53,12 +67,21 @@ module Decidim
           expect(serialized).to include(body: proposal.body)
         end
 
+        it "serializes the origin" do
+          expect(serialized).to include(collaborative_draft_origin: proposal.collaborative_draft_origin)
+        end
+
         it "serializes the amount of supports" do
           expect(serialized).to include(supports: proposal.proposal_votes_count)
         end
 
         it "serializes the amount of comments" do
           expect(serialized).to include(comments: proposal.comments.count)
+        end
+
+        it "serializes the amount of amendments" do
+          expect(serialized).to include(amendments: proposal.amendments.count)
+          expect(serialized[:amendments]).to eq(1)
         end
 
         it "serializes the date of creation" do
@@ -91,6 +114,11 @@ module Decidim
           expect(serialized).to include(reference: proposal.reference)
         end
 
+        it "serializes attachments url" do
+          expect(serialized[:attachments_url].first).to eq(proposal.organization.host + proposal.attachments.first.url)
+          expect(serialized[:attachments_url].length).to eq(1)
+        end
+
         it "serializes the amount of attachments" do
           expect(serialized).to include(attachments: proposal.attachments.count)
         end
@@ -102,6 +130,35 @@ module Decidim
         it "serializes related proposals" do
           expect(serialized[:related_proposals].length).to eq(2)
           expect(serialized[:related_proposals].first).to match(%r{http.*/proposals})
+        end
+
+        it "serializes the author profile url" do
+          expect(serialized[:authors_url]).to all(match(%r{http.*/profiles}))
+          expect(serialized[:authors_url].length).to eq(1)
+        end
+
+        context "when multiple authors" do
+          let!(:users) { create_list(:user, 3, organization: proposals_component.organization) }
+          let!(:proposal) { create(:proposal, :accepted) }
+
+          before do
+            users.each { |user| proposal.add_coauthor(user) }
+          end
+
+          it "serializes the authors url" do
+            expect(serialized[:authors_url]).to all(match(%r{http.*/profiles}))
+            expect(serialized[:authors_url].length).to eq(4)
+          end
+
+          context "when authors includes an official" do
+            let!(:proposal) { create(:proposal, :accepted, :official) }
+
+            it "serializes the authors" do
+              expect(serialized[:authors_url].first).to eq("")
+              expect(serialized[:authors_url][1..-1]).to all(match(%r{http.*/profiles}))
+              expect(serialized[:authors_url].length).to eq(4)
+            end
+          end
         end
       end
     end